refactor: Move get current problem LC ID to reusable helpers
